pyproject.toml
[project]
name = "datahack-nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "whisperx>=3.8.5",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "soundfile>=0.12.1",
    "praat-parselmouth>=0.4.3",
    "transformers>=4.40.0",
    "huggingface-hub>=0.23.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
]
